COC-172-refactor(#97): useSpace 훅 분리
