+++ src/constants/queryKeys.ts
@@ -1,6 +1,8 @@
 const QUERY_KEYS = {
   STUDY_DETAIL: 'study_detail',
   STUDY_LIST: 'study_list',
+  SPACE_DETAIL: 'space_detail',
+  SPACE_TAB: 'space_tab',
 } as const;
 
 export default QUERY_KEYS;
